requirements.txt
asgiref==3.8.1
certifi==2024.6.2
charset-normalizer==3.3.2
Django==5.0.6
django-dotenv==1.4.2
idna==3.7
PyJWT==2.8.0
requests==2.32.3
sqlparse==0.5.0
urllib3==2.2.2
